@@ -1,6 +1,6 @@
 {
   "name": "mern-marketplace",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "A MERN stack based online marketplace application",
   "author": "Shama Hoque",
   "license": "MIT",
@@ -23,45 +23,46 @@
     "start": "NODE_ENV=production node ./dist/server.generated.js"
   },
   "engines": {
-    "node": "8.11.1",
-    "npm": "5.8.0"
+    "node": "13.12.0",
+    "npm": "6.14.4"
   },
   "devDependencies": {
-    "babel-core": "^6.26.2",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "file-loader": "^1.1.11",
-    "nodemon": "^1.17.3",
-    "webpack": "^4.6.0",
-    "webpack-cli": "^2.0.15",
-    "webpack-dev-middleware": "^3.1.3",
-    "webpack-hot-middleware": "^2.22.1",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
+    "babel-loader": "^8.1.0",
+    "file-loader": "^6.0.0",
+    "nodemon": "^2.0.2",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-hot-middleware": "^2.25.0",
     "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
-    "body-parser": "^1.18.2",
-    "compression": "^1.7.2",
-    "cookie-parser": "^1.4.3",
-    "cors": "^2.8.4",
-    "express": "^4.16.3",
+    "@hot-loader/react-dom": "^16.13.0",
+    "@material-ui/core": "^4.9.8",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/styles": "^4.9.6",
+    "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
+    "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
     "express-jwt": "^5.3.1",
-    "formidable": "^1.2.1",
-    "helmet": "^3.12.0",
-    "jsonwebtoken": "^8.2.1",
-    "lodash": "^4.17.10",
-    "material-ui": "^1.0.0-beta.43",
-    "material-ui-icons": "^1.0.0-beta.36",
-    "mongoose": "^5.0.16",
-    "query-string": "^6.0.0",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-hot-loader": "^4.1.2",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-stripe-elements": "^1.6.0",
-    "request": "^2.85.0",
-    "stripe": "^5.8.0"
+    "formidable": "^1.2.2",
+    "helmet": "^3.22.0",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.15",
+    "mongoose": "^5.9.7",
+    "query-string": "^6.11.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-hot-loader": "^4.12.20",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "react-stripe-elements": "^6.1.1",
+    "request": "^2.88.2",
+    "stripe": "^8.38.0"
   }
 }
