@@ -1,16 +1,15 @@
-import React, {Component} from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Card from 'material-ui/Card'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
+import React, {useState, useEffect} from 'react'
+import Card from '@material-ui/core/Card'
+import {makeStyles} from '@material-ui/core/styles'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
 import auth from './../auth/auth-helper'
 import cart from './cart-helper.js'
 import PlaceOrder from './PlaceOrder'
 import {Elements} from 'react-stripe-elements'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   card: {
     margin: '24px 0px',
     padding: '16px 40px 90px 40px',
@@ -26,86 +25,72 @@ const styles = theme => ({
   },
   addressField: {
     marginTop: "4px",
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: "45%"
   },
   streetField: {
     marginTop: "4px",
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: "93%"
   },
   textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: "90%"
   }
-})
+}))
 
-class Checkout extends Component {
-  state = {
+export default function Checkout (){
+  const classes = useStyles()
+  const user = auth.isAuthenticated().user
+  const [values, setValues] = useState({
     checkoutDetails: {
-      customer_name: '',
-      customer_email:'',
+      products: cart.getCart(),
+      customer_name: user.name,
+      customer_email:user.email,
       delivery_address: { street: '', city: '', state: '', zipcode: '', country:''}
     },
     error: ''
-  }
-  componentDidMount = () => {
-    let user = auth.isAuthenticated().user
-    let checkoutDetails = this.state.checkoutDetails
-    checkoutDetails.products = cart.getCart()
-    checkoutDetails.customer_name = user.name
-    checkoutDetails.customer_email = user.email
-    this.setState({checkoutDetails: checkoutDetails})
-  }
+  })
 
-  handleCustomerChange = name => event => {
-    let checkoutDetails = this.state.checkoutDetails
+  const handleCustomerChange = name => event => {
+    let checkoutDetails = values.checkoutDetails
     checkoutDetails[name] = event.target.value || undefined
-    this.setState({checkoutDetails: checkoutDetails})
+    setValues({...values, checkoutDetails: checkoutDetails})
   }
 
-  handleAddressChange = name => event => {
-    let checkoutDetails = this.state.checkoutDetails
+  const handleAddressChange = name => event => {
+    let checkoutDetails = values.checkoutDetails
     checkoutDetails.delivery_address[name] = event.target.value || undefined
-    this.setState({checkoutDetails: checkoutDetails})
+    setValues({...values, checkoutDetails: checkoutDetails})
   }
 
-  render() {
-    const {classes} = this.props
     return (
       <Card className={classes.card}>
         <Typography type="title" className={classes.title}>
           Checkout
         </Typography>
-        <TextField id="name" label="Name" className={classes.textField} value={this.state.checkoutDetails.customer_name} onChange={this.handleCustomerChange('customer_name')} margin="normal"/><br/>
-        <TextField id="email" type="email" label="Email" className={classes.textField} value={this.state.checkoutDetails.customer_email} onChange={this.handleCustomerChange('customer_email')} margin="normal"/><br/>
+        <TextField id="name" label="Name" className={classes.textField} value={values.checkoutDetails.customer_name} onChange={handleCustomerChange('customer_name')} margin="normal"/><br/>
+        <TextField id="email" type="email" label="Email" className={classes.textField} value={values.checkoutDetails.customer_email} onChange={handleCustomerChange('customer_email')} margin="normal"/><br/>
         <Typography type="subheading" component="h3" className={classes.subheading}>
             Delivery Address
         </Typography>
-        <TextField id="street" label="Street Address" className={classes.streetField} value={this.state.checkoutDetails.delivery_address.street} onChange={this.handleAddressChange('street')} margin="normal"/><br/>
-        <TextField id="city" label="City" className={classes.addressField} value={this.state.checkoutDetails.delivery_address.city} onChange={this.handleAddressChange('city')} margin="normal"/>
-        <TextField id="state" label="State" className={classes.addressField} value={this.state.checkoutDetails.delivery_address.state} onChange={this.handleAddressChange('state')} margin="normal"/><br/>
-        <TextField id="zipcode" label="Zip Code" className={classes.addressField} value={this.state.checkoutDetails.delivery_address.zipcode} onChange={this.handleAddressChange('zipcode')} margin="normal"/>
-        <TextField id="country" label="Country" className={classes.addressField} value={this.state.checkoutDetails.delivery_address.country} onChange={this.handleAddressChange('country')} margin="normal"/>
+        <TextField id="street" label="Street Address" className={classes.streetField} value={values.checkoutDetails.delivery_address.street} onChange={handleAddressChange('street')} margin="normal"/><br/>
+        <TextField id="city" label="City" className={classes.addressField} value={values.checkoutDetails.delivery_address.city} onChange={handleAddressChange('city')} margin="normal"/>
+        <TextField id="state" label="State" className={classes.addressField} value={values.checkoutDetails.delivery_address.state} onChange={handleAddressChange('state')} margin="normal"/><br/>
+        <TextField id="zipcode" label="Zip Code" className={classes.addressField} value={values.checkoutDetails.delivery_address.zipcode} onChange={handleAddressChange('zipcode')} margin="normal"/>
+        <TextField id="country" label="Country" className={classes.addressField} value={values.checkoutDetails.delivery_address.country} onChange={handleAddressChange('country')} margin="normal"/>
         <br/> {
-            this.state.error && (<Typography component="p" color="error">
+            values.error && (<Typography component="p" color="error">
                 <Icon color="error" className={classes.error}>error</Icon>
-                {this.state.error}</Typography>)
+                {values.error}</Typography>)
           }
         <div>
           <Elements>
-            <PlaceOrder checkoutDetails={this.state.checkoutDetails} />
+            <PlaceOrder checkoutDetails={values.checkoutDetails} />
           </Elements>
         </div>
       </Card>)
-  }
 }
-
-Checkout.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(Checkout)
