@@ -1,61 +1,68 @@
-import React, {Component} from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Grid from 'material-ui/Grid'
+import React, {useState, useEffect} from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import Grid from '@material-ui/core/Grid'
 import Suggestions from './../product/Suggestions'
 import {listLatest, listCategories} from './../product/api-product.js'
 import Search from './../product/Search'
 import Categories from './../product/Categories'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
     margin: 30,
   }
-})
+}))
 
-class Home extends Component {
-  state={
-    suggestionTitle: "Latest Products",
-    suggestions: [],
-    categories: []
-  }
-  componentDidMount = () => {
-    listLatest().then((data) => {
+
+export default function Home(){
+  const classes = useStyles();
+  const [suggestionTitle, setSuggestionTitle] = useState("Latest Products")
+  const [categories, setCategories] = useState([])
+  const [suggestions, setSuggestions] = useState([])
+  
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+    listLatest(signal).then((data) => {
       if (data.error) {
         console.log(data.error)
       } else {
-        this.setState({suggestions: data})
+        setSuggestions(data)
       }
     })
-    listCategories().then((data) => {
+    return function cleanup(){
+      abortController.abort()
+    }
+  }, [])
+
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+    listCategories(signal).then((data) => {
       if (data.error) {
         console.log(data.error)
       } else {
-        this.setState({categories: data})
+        setCategories(data)
       }
     })
-  }
-  render() {
-    const {classes} = this.props
+    return function cleanup(){
+      abortController.abort()
+    }
+  }, [])
+
     return (
       <div className={classes.root}>
-        <Grid container spacing={24}>
+        <Grid container spacing={2}>
           <Grid item xs={8} sm={8}>
-            <Search categories={this.state.categories}/>
-            <Categories categories={this.state.categories}/>
+            <Search categories={categories}/>
+            <Categories categories={categories}/>
           </Grid>
           <Grid item xs={4} sm={4}>
-            <Suggestions products={this.state.suggestions} title={this.state.suggestionTitle}/>
+            <Suggestions products={suggestions} title={suggestionTitle}/>
           </Grid>
         </Grid>
       </div>
     )
-  }
 }
 
-Home.propTypes = {
-  classes: PropTypes.object.isRequired
-}
 
-export default withStyles(styles)(Home)
