@@ -1,64 +1,65 @@
-import React, {Component} from 'react'
+import React, {useState} from 'react'
 import PropTypes from 'prop-types'
-import IconButton from 'material-ui/IconButton'
-import Button from 'material-ui/Button'
-import DeleteIcon from 'material-ui-icons/Delete'
-import Dialog, {DialogActions, DialogContent, DialogContentText, DialogTitle} from 'material-ui/Dialog'
+import IconButton from '@material-ui/core/IconButton'
+import Button from '@material-ui/core/Button'
+import DeleteIcon from '@material-ui/icons/Delete'
+import Dialog from '@material-ui/core/Dialog'
+import DialogActions from '@material-ui/core/DialogActions'
+import DialogContent from '@material-ui/core/DialogContent'
+import DialogContentText from '@material-ui/core/DialogContentText'
+import DialogTitle from '@material-ui/core/DialogTitle'
 import auth from './../auth/auth-helper'
 import {remove} from './api-product.js'
 
-class DeleteProduct extends Component {
-  state = {
-    open: false
+export default function DeleteProduct(props) {
+  const [open, setOpen] = useState(false)
+  
+  const jwt = auth.isAuthenticated()
+  const clickButton = () => {
+    setOpen(true)
   }
-  clickButton = () => {
-    this.setState({open: true})
-  }
-  deleteProduct = () => {
-    const jwt = auth.isAuthenticated()
+  const deleteProduct = () => {
     remove({
-      shopId: this.props.shopId,
-      productId: this.props.product._id
+      shopId: props.shopId,
+      productId: props.product._id
     }, {t: jwt.token}).then((data) => {
       if (data.error) {
         console.log(data.error)
       } else {
-        this.setState({open: false}, () => {
-          this.props.onRemove(this.props.product)
-        })
+        setOpen(false)
+        props.onRemove(props.product)
       }
     })
   }
-  handleRequestClose = () => {
-    this.setState({open: false})
+  const handleRequestClose = () => {
+    setOpen(false)
   }
-  render() {
     return (<span>
-      <IconButton aria-label="Delete" onClick={this.clickButton} color="secondary">
+      <IconButton aria-label="Delete" onClick={clickButton} color="secondary">
         <DeleteIcon/>
       </IconButton>
-      <Dialog open={this.state.open} onClose={this.handleRequestClose}>
-        <DialogTitle>{"Delete "+this.props.product.name}</DialogTitle>
+      <Dialog open={open} onClose={handleRequestClose}>
+        <DialogTitle>{"Delete "+props.product.name}</DialogTitle>
         <DialogContent>
           <DialogContentText>
-            Confirm to delete your product {this.props.product.name}.
+            Confirm to delete your product {props.product.name}.
           </DialogContentText>
         </DialogContent>
         <DialogActions>
-          <Button onClick={this.handleRequestClose} color="primary">
+          <Button onClick={handleRequestClose} color="primary">
             Cancel
           </Button>
-          <Button onClick={this.deleteProduct} color="secondary" autoFocus="autoFocus">
+          <Button onClick={deleteProduct} color="secondary" autoFocus="autoFocus">
             Confirm
           </Button>
         </DialogActions>
       </Dialog>
     </span>)
-  }
+
 }
 DeleteProduct.propTypes = {
   shopId: PropTypes.string.isRequired,
   product: PropTypes.object.isRequired,
   onRemove: PropTypes.func.isRequired
 }
-export default DeleteProduct
+
