@@ -1,28 +1,30 @@
-import React, {Component} from 'react'
+import React, {useEffect, useState} from 'react'
+import Card from '@material-ui/core/Card'
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent'
+import Button from '@material-ui/core/Button'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
+import Avatar from '@material-ui/core/Avatar'
 import auth from './../auth/auth-helper'
-import Card, {CardActions, CardContent} from 'material-ui/Card'
-import Button from 'material-ui/Button'
-import FileUpload from 'material-ui-icons/FileUpload'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
-import Avatar from 'material-ui/Avatar'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import FileUpload from '@material-ui/icons/AddPhotoAlternate'
+import { makeStyles } from '@material-ui/core/styles'
+import {withStyles} from '@material-ui/core/styles'
 import {read, update} from './api-product.js'
 import {Link, Redirect} from 'react-router-dom'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   card: {
     margin: 'auto',
     textAlign: 'center',
-    marginTop: theme.spacing.unit * 3,
-    marginBottom: theme.spacing.unit * 2,
+    marginTop: theme.spacing(3),
+    marginBottom: theme.spacing(2),
     maxWidth: 500,
-    paddingBottom: theme.spacing.unit * 2
+    paddingBottom: theme.spacing(2)
   },
   title: {
-    margin: theme.spacing.unit * 2,
+    margin: theme.spacing(2),
     color: theme.palette.protectedTitle,
     fontSize: '1.2em'
   },
@@ -30,13 +32,13 @@ const styles = theme => ({
     verticalAlign: 'middle'
   },
   textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: 400
   },
   submit: {
     margin: 'auto',
-    marginBottom: theme.spacing.unit * 2
+    marginBottom: theme.spacing(2)
   },
   bigAvatar: {
     width: 60,
@@ -49,12 +51,11 @@ const styles = theme => ({
   filename:{
     marginLeft:'10px'
   }
-})
+}))
 
-class EditProduct extends Component {
-  constructor({match}) {
-    super()
-    this.state = {
+export default function EditProduct ({match}) {
+  const classes = useStyles()
+  const [values, setValues] = useState({
       name: '',
       description: '',
       image: '',
@@ -63,53 +64,59 @@ class EditProduct extends Component {
       price: '',
       redirect: false,
       error: ''
-    }
-    this.match = match
-  }
-
-  componentDidMount = () => {
-    this.productData = new FormData()
-    read({
-      productId: this.match.params.productId
-    }).then((data) => {
-      if (data.error) {
-        this.setState({error: data.error})
-      } else {
-        this.setState({id: data._id, name: data.name, description: data.description, category: data.category, quantity:data.quantity, price: data.price})
-      }
     })
-  }
-  clickSubmit = () => {
+
     const jwt = auth.isAuthenticated()
+    useEffect(() => {
+      const abortController = new AbortController()
+      const signal = abortController.signal
+      read({
+        productId: match.params.productId
+      }, signal).then((data) => {
+        if (data.error) {
+          setValues({...values, error: data.error})
+        } else {
+          setValues({...values, id: data._id, name: data.name, description: data.description, category: data.category, quantity:data.quantity, price: data.price})
+        }
+      })
+    return function cleanup(){
+      abortController.abort()
+    }
+  }, [])
+  const clickSubmit = () => {
+    let productData = new FormData()
+    values.name && productData.append('name', values.name)
+    values.description && productData.append('description', values.description)
+    values.image && productData.append('image', values.image)
+    values.category && productData.append('category', values.category)
+    values.quantity && productData.append('quantity', values.quantity)
+    values.price && productData.append('price', values.price)
+  
     update({
-      shopId: this.match.params.shopId,
-      productId: this.match.params.productId
+      shopId: match.params.shopId,
+      productId: match.params.productId
     }, {
       t: jwt.token
-    }, this.productData).then((data) => {
+    }, productData).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({...values, error: data.error})
       } else {
-        this.setState({'redirect': true})
+        setValues({...values, 'redirect': true})
       }
     })
   }
-  handleChange = name => event => {
+  const handleChange = name => event => {
     const value = name === 'image'
       ? event.target.files[0]
       : event.target.value
-    this.productData.set(name, value)
-    this.setState({ [name]: value })
+    setValues({...values,  [name]: value })
   }
-
-  render() {
-    const imageUrl = this.state.id
-          ? `/api/product/image/${this.state.id}?${new Date().getTime()}`
+    const imageUrl = values.id
+          ? `/api/product/image/${values.id}?${new Date().getTime()}`
           : '/api/product/defaultphoto'
-    if (this.state.redirect) {
-      return (<Redirect to={'/seller/shop/edit/'+this.match.params.shopId}/>)
+    if (values.redirect) {
+      return (<Redirect to={'/seller/shop/edit/'+match.params.shopId}/>)
     }
-    const {classes} = this.props
     return (<div>
       <Card className={classes.card}>
         <CardContent>
@@ -117,44 +124,37 @@ class EditProduct extends Component {
             Edit Product
           </Typography><br/>
           <Avatar src={imageUrl} className={classes.bigAvatar}/><br/>
-          <input accept="image/*" onChange={this.handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
+          <input accept="image/*" onChange={handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
           <label htmlFor="icon-button-file">
-            <Button variant="raised" color="secondary" component="span">
+            <Button variant="contained" color="secondary" component="span">
               Change Image
               <FileUpload/>
             </Button>
-          </label> <span className={classes.filename}>{this.state.image ? this.state.image.name : ''}</span><br/>
-          <TextField id="name" label="Name" className={classes.textField} value={this.state.name} onChange={this.handleChange('name')} margin="normal"/><br/>
+          </label> <span className={classes.filename}>{values.image ? values.image.name : ''}</span><br/>
+          <TextField id="name" label="Name" className={classes.textField} value={values.name} onChange={handleChange('name')} margin="normal"/><br/>
           <TextField
             id="multiline-flexible"
             label="Description"
             multiline
             rows="3"
-            value={this.state.description}
-            onChange={this.handleChange('description')}
+            value={values.description}
+            onChange={handleChange('description')}
             className={classes.textField}
             margin="normal"
           /><br/>
-          <TextField id="category" label="Category" className={classes.textField} value={this.state.category} onChange={this.handleChange('category')} margin="normal"/><br/>
-          <TextField id="quantity" label="Quantity" className={classes.textField} value={this.state.quantity} onChange={this.handleChange('quantity')} type="number" margin="normal"/><br/>
-          <TextField id="price" label="Price" className={classes.textField} value={this.state.price} onChange={this.handleChange('price')} type="number" margin="normal"/><br/>
+          <TextField id="category" label="Category" className={classes.textField} value={values.category} onChange={handleChange('category')} margin="normal"/><br/>
+          <TextField id="quantity" label="Quantity" className={classes.textField} value={values.quantity} onChange={handleChange('quantity')} type="number" margin="normal"/><br/>
+          <TextField id="price" label="Price" className={classes.textField} value={values.price} onChange={handleChange('price')} type="number" margin="normal"/><br/>
           {
-            this.state.error && (<Typography component="p" color="error">
+            values.error && (<Typography component="p" color="error">
               <Icon color="error" className={classes.error}>error</Icon>
-              {this.state.error}</Typography>)
+              {values.error}</Typography>)
           }
         </CardContent>
         <CardActions>
-          <Button color="primary" variant="raised" onClick={this.clickSubmit} className={classes.submit}>Update</Button>
-          <Link to={'/seller/shops/edit/'+this.match.params.shopId} className={classes.submit}><Button variant="raised">Cancel</Button></Link>
+          <Button color="primary" variant="contained" onClick={clickSubmit} className={classes.submit}>Update</Button>
+          <Link to={'/seller/shops/edit/'+match.params.shopId} className={classes.submit}><Button variant="contained">Cancel</Button></Link>
         </CardActions>
       </Card>
     </div>)
-  }
 }
-
-EditProduct.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(EditProduct)
