@@ -1,16 +1,17 @@
-import React, {Component} from 'react'
-import Card, {CardHeader, CardMedia} from 'material-ui/Card'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
-import Grid from 'material-ui/Grid'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import React, {useState, useEffect}  from 'react'
+import Card from '@material-ui/core/Card'
+import CardHeader from '@material-ui/core/CardHeader'
+import CardMedia from '@material-ui/core/CardMedia'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
+import Grid from '@material-ui/core/Grid'
+import {makeStyles} from '@material-ui/core/styles'
 import {read, listRelated} from './api-product.js'
 import {Link} from 'react-router-dom'
 import Suggestions from './../product/Suggestions'
 import AddToCart from './../cart/AddToCart'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
     margin: 30,
@@ -57,58 +58,60 @@ const styles = theme => ({
     margin: '8px 24px',
     display: 'inline-block'
   }
-})
+}))
 
-class Product extends Component {
-  constructor({match}) {
-    super()
-    this.state = {
-      product: {shop: {}},
-      suggestions: [],
-      suggestionTitle: 'Related Products'
+export default function Product ({match}) {
+  const classes = useStyles()
+  const [product, setProduct] = useState({shop:{}})
+  const [suggestions, setSuggestions] = useState([])
+  const [error, setError] = useState('')
+    useEffect(() => {
+      const abortController = new AbortController()
+      const signal = abortController.signal
+  
+      read({productId: match.params.productId}, signal).then((data) => {
+        if (data.error) {
+          setError(data.error)
+        } else {
+          setProduct(data)
+        }
+      })
+    return function cleanup(){
+      abortController.abort()
     }
-    this.match = match
-  }
-  loadProduct = (productId) => {
-    read({productId: productId}).then((data) => {
-      if (data.error) {
-        this.setState({error: data.error})
-      } else {
-        this.setState({product: data})
+  }, [match.params.productId])
+
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
         listRelated({
-          productId: data._id}).then((data) => {
+          productId: match.params.productId}, signal).then((data) => {
           if (data.error) {
-            console.log(data.error)
+            setError(data.error)
           } else {
-            this.setState({suggestions: data})
+            setSuggestions(data)
           }
         })
-     }
-    })
+  return function cleanup(){
+    abortController.abort()
   }
-  componentDidMount = () => {
-    this.loadProduct(this.match.params.productId)
-  }
-  componentWillReceiveProps = (props) => {
-    this.loadProduct(props.match.params.productId)
-  }
+}, [match.params.productId])
 
-  render() {
-    const imageUrl = this.state.product._id
-          ? `/api/product/image/${this.state.product._id}?${new Date().getTime()}`
+    const imageUrl = product._id
+          ? `/api/product/image/${product._id}?${new Date().getTime()}`
           : '/api/product/defaultphoto'
-    const {classes} = this.props
     return (
         <div className={classes.root}>
-          <Grid container spacing={40}>
+          <Grid container spacing={10}>
             <Grid item xs={7} sm={7}>
               <Card className={classes.card}>
                 <CardHeader
-                  title={this.state.product.name}
-                  subheader={this.state.product.quantity > 0? 'In Stock': 'Out of Stock'}
+                  title={product.name}
+                  subheader={product.quantity > 0? 'In Stock': 'Out of Stock'}
                   action={
                     <span className={classes.action}>
-                      <AddToCart cartStyle={classes.addCart} item={this.state.product}/>
+                      <AddToCart cartStyle={classes.addCart} item={product}/>
                     </span>
                   }
                 />
@@ -116,14 +119,14 @@ class Product extends Component {
                   <CardMedia
                     className={classes.media}
                     image={imageUrl}
-                    title={this.state.product.name}
+                    title={product.name}
                   />
-                  <Typography component="p" type="subheading" className={classes.subheading}>
-                    {this.state.product.description}<br/>
-                    <span className={classes.price}>$ {this.state.product.price}</span>
-                    <Link to={'/shops/'+this.state.product.shop._id} className={classes.link}>
+                  <Typography component="p" variant="subtitle1" className={classes.subheading}>
+                    {product.description}<br/>
+                    <span className={classes.price}>$ {product.price}</span>
+                    <Link to={'/shops/'+product.shop._id} className={classes.link}>
                       <span>
-                        <Icon className={classes.icon}>shopping_basket</Icon> {this.state.product.shop.name}
+                        <Icon className={classes.icon}>shopping_basket</Icon> {product.shop.name}
                       </span>
                     </Link>
                   </Typography>
@@ -131,17 +134,10 @@ class Product extends Component {
                 </div>
               </Card>
             </Grid>
-            {this.state.suggestions.length > 0 &&
+            {suggestions.length > 0 &&
               (<Grid item xs={5} sm={5}>
-                <Suggestions  products={this.state.suggestions} title='Related Products'/>
+                <Suggestions  products={suggestions} title='Related Products'/>
               </Grid>)}
           </Grid>
         </div>)
-  }
 }
-
-Product.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(Product)
