@@ -1,48 +1,49 @@
-import React, {Component} from 'react'
+import React, {useEffect, useState} from 'react'
+import Card from '@material-ui/core/Card'
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent'
+import Button from '@material-ui/core/Button'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
+import Avatar from '@material-ui/core/Avatar'
 import auth from './../auth/auth-helper'
-import Card, {CardActions, CardContent, CardMedia} from 'material-ui/Card'
-import Button from 'material-ui/Button'
-import FileUpload from 'material-ui-icons/FileUpload'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
-import Avatar from 'material-ui/Avatar'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import FileUpload from '@material-ui/icons/AddPhotoAlternate'
+import { makeStyles } from '@material-ui/core/styles'
 import {read, update} from './api-shop.js'
 import {Redirect} from 'react-router-dom'
-import Grid from 'material-ui/Grid'
+import Grid from '@material-ui/core/Grid'
 import MyProducts from './../product/MyProducts'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
     margin: 30,
   },
   card: {
     textAlign: 'center',
-    paddingBottom: theme.spacing.unit * 2
+    paddingBottom: theme.spacing(2)
   },
   title: {
-    margin: theme.spacing.unit * 2,
+    margin: theme.spacing(2),
     color: theme.palette.protectedTitle,
     fontSize: '1.2em'
   },
   subheading: {
-    marginTop: theme.spacing.unit * 2,
+    marginTop: theme.spacing(2),
     color: theme.palette.openTitle
   },
   error: {
     verticalAlign: 'middle'
   },
   textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: 400
   },
   submit: {
     margin: 'auto',
-    marginBottom: theme.spacing.unit * 2
+    marginBottom: theme.spacing(2)
   },
   bigAvatar: {
     width: 60,
@@ -55,66 +56,68 @@ const styles = theme => ({
   filename:{
     marginLeft:'10px'
   }
-})
+}))
 
-class EditShop extends Component {
-  constructor({match}) {
-    super()
-    this.state = {
+export default function EditShop ({match}) {
+  const classes = useStyles()
+  const [values, setValues] = useState({
       name: '',
       description: '',
       image: '',
       redirect: false,
-      error: ''
-    }
-    this.match = match
-  }
-
-  componentDidMount = () => {
-    this.shopData = new FormData()
-    const jwt = auth.isAuthenticated()
+      error: '',
+      id: ''
+  })
+  const jwt = auth.isAuthenticated()
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
     read({
-      shopId: this.match.params.shopId
-    }, {t: jwt.token}).then((data) => {
+      shopId: match.params.shopId
+    }, signal).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({...values, error: data.error})
       } else {
-        this.setState({id: data._id, name: data.name, description: data.description, owner: data.owner.name})
+        setValues({...values, id: data._id, name: data.name, description: data.description, owner: data.owner.name})
       }
     })
-  }
-  clickSubmit = () => {
-    const jwt = auth.isAuthenticated()
+    return function cleanup(){
+      abortController.abort()
+    }
+  }, [])
+
+  const clickSubmit = () => {
+    let shopData = new FormData()
+    values.name && shopData.append('name', values.name)
+    values.description && shopData.append('description', values.description)
+    values.image && shopData.append('image', values.image)
     update({
-      shopId: this.match.params.shopId
+      shopId: match.params.shopId
     }, {
       t: jwt.token
-    }, this.shopData).then((data) => {
+    }, shopData).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({...values, error: data.error})
       } else {
-        this.setState({'redirect': true})
+        setValues({...values, 'redirect': true})
       }
     })
   }
-  handleChange = name => event => {
+  const handleChange = name => event => {
     const value = name === 'image'
       ? event.target.files[0]
       : event.target.value
-    this.shopData.set(name, value)
-    this.setState({ [name]: value })
+    setValues({...values,  [name]: value })
   }
 
-  render() {
-    const logoUrl = this.state.id
-          ? `/api/shops/logo/${this.state.id}?${new Date().getTime()}`
+    const logoUrl = values.id
+          ? `/api/shops/logo/${values.id}?${new Date().getTime()}`
           : '/api/shops/defaultphoto'
-    if (this.state.redirect) {
+    if (values.redirect) {
       return (<Redirect to={'/seller/shops'}/>)
     }
-    const {classes} = this.props
     return (<div className={classes.root}>
-      <Grid container spacing={24}>
+      <Grid container spacing={8}>
         <Grid item xs={6} sm={6}>
           <Card className={classes.card}>
             <CardContent>
@@ -123,49 +126,42 @@ class EditShop extends Component {
               </Typography>
               <br/>
               <Avatar src={logoUrl} className={classes.bigAvatar}/><br/>
-              <input accept="image/*" onChange={this.handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
+              <input accept="image/*" onChange={handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
               <label htmlFor="icon-button-file">
-                <Button variant="raised" color="default" component="span">
+                <Button variant="contained" color="default" component="span">
                   Change Logo
                   <FileUpload/>
                 </Button>
-              </label> <span className={classes.filename}>{this.state.image ? this.state.image.name : ''}</span><br/>
-              <TextField id="name" label="Name" className={classes.textField} value={this.state.name} onChange={this.handleChange('name')} margin="normal"/><br/>
+              </label> <span className={classes.filename}>{values.image ? values.image.name : ''}</span><br/>
+              <TextField id="name" label="Name" className={classes.textField} value={values.name} onChange={handleChange('name')} margin="normal"/><br/>
               <TextField
                 id="multiline-flexible"
                 label="Description"
                 multiline
                 rows="3"
-                value={this.state.description}
-                onChange={this.handleChange('description')}
+                value={values.description}
+                onChange={handleChange('description')}
                 className={classes.textField}
                 margin="normal"
               /><br/>
               <Typography type="subheading" component="h4" className={classes.subheading}>
-                Owner: {this.state.owner}
+                Owner: {values.owner}
               </Typography><br/>
               {
-                this.state.error && (<Typography component="p" color="error">
+                values.error && (<Typography component="p" color="error">
                     <Icon color="error" className={classes.error}>error</Icon>
-                    {this.state.error}
+                    {values.error}
                   </Typography>)
               }
             </CardContent>
             <CardActions>
-              <Button color="primary" variant="raised" onClick={this.clickSubmit} className={classes.submit}>Update</Button>
+              <Button color="primary" variant="contained" onClick={clickSubmit} className={classes.submit}>Update</Button>
             </CardActions>
           </Card>
           </Grid>
           <Grid item xs={6} sm={6}>
-            <MyProducts shopId={this.match.params.shopId}/>
+            <MyProducts shopId={match.params.shopId}/>
           </Grid>
         </Grid>
     </div>)
-  }
 }
-
-EditShop.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(EditShop)
