@@ -1,39 +1,41 @@
-import React, {Component} from 'react'
-import Card, {CardActions, CardContent} from 'material-ui/Card'
-import Button from 'material-ui/Button'
-import FileUpload from 'material-ui-icons/FileUpload'
+import React, {useState} from 'react'
+import Card from '@material-ui/core/Card'
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent'
+import Button from '@material-ui/core/Button'
+import FileUpload from '@material-ui/icons/AddPhotoAlternate'
 import auth from './../auth/auth-helper'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
+import { makeStyles } from '@material-ui/core/styles'
 import {create} from './api-shop.js'
 import {Link, Redirect} from 'react-router-dom'
-const styles = theme => ({
+
+const useStyles = makeStyles(theme => ({
   card: {
     maxWidth: 600,
     margin: 'auto',
     textAlign: 'center',
-    marginTop: theme.spacing.unit * 5,
-    paddingBottom: theme.spacing.unit * 2
+    marginTop: theme.spacing(5),
+    paddingBottom: theme.spacing(2)
   },
   error: {
     verticalAlign: 'middle'
   },
   title: {
-    marginTop: theme.spacing.unit * 2,
+    marginTop: theme.spacing(2),
     color: theme.palette.openTitle,
     fontSize: '1em'
   },
   textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: 300
   },
   submit: {
     margin: 'auto',
-    marginBottom: theme.spacing.unit * 2
+    marginBottom: theme.spacing(2)
   },
   input: {
     display: 'none'
@@ -41,46 +43,46 @@ const styles = theme => ({
   filename:{
     marginLeft:'10px'
   }
-})
+}))
 
-class NewShop extends Component {
-  state = {
+export default function NewShop() {
+  const classes = useStyles();
+  const [values, setValues] = useState({
       name: '',
       description: '',
       image: '',
       redirect: false,
       error: ''
-  }
-  componentDidMount = () => {
-    this.shopData = new FormData()
-  }
-  handleChange = name => event => {
+  })
+  const jwt = auth.isAuthenticated()
+
+  const handleChange = name => event => {
     const value = name === 'image'
       ? event.target.files[0]
       : event.target.value
-    this.shopData.set(name, value)
-    this.setState({ [name]: value })
+    setValues({...values, [name]: value })
   }
-  clickSubmit = () => {
-    const jwt = auth.isAuthenticated()
+  const clickSubmit = () => {
+    let shopData = new FormData()
+    values.name && shopData.append('name', values.name)
+    values.description && shopData.append('description', values.description)
+    values.image && shopData.append('image', values.image)
     create({
       userId: jwt.user._id
     }, {
       t: jwt.token
-    }, this.shopData).then((data) => {
+    }, shopData).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({...values, error: data.error})
       } else {
-        this.setState({error: '', redirect: true})
+        setValues({...values, error: '', redirect: true})
       }
     })
   }
 
-  render() {
-    if (this.state.redirect) {
+    if (values.redirect) {
       return (<Redirect to={'/seller/shops'}/>)
     }
-    const {classes} = this.props
     return (<div>
       <Card className={classes.card}>
         <CardContent>
@@ -88,40 +90,33 @@ class NewShop extends Component {
             New Shop
           </Typography>
           <br/>
-          <input accept="image/*" onChange={this.handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
+          <input accept="image/*" onChange={handleChange('image')} className={classes.input} id="icon-button-file" type="file" />
           <label htmlFor="icon-button-file">
-            <Button variant="raised" color="secondary" component="span">
+            <Button variant="contained" color="secondary" component="span">
               Upload Logo
               <FileUpload/>
             </Button>
-          </label> <span className={classes.filename}>{this.state.image ? this.state.image.name : ''}</span><br/>
-          <TextField id="name" label="Name" className={classes.textField} value={this.state.name} onChange={this.handleChange('name')} margin="normal"/><br/>
+          </label> <span className={classes.filename}>{values.image ? values.image.name : ''}</span><br/>
+          <TextField id="name" label="Name" className={classes.textField} value={values.name} onChange={handleChange('name')} margin="normal"/><br/>
           <TextField
             id="multiline-flexible"
             label="Description"
             multiline
             rows="2"
-            value={this.state.description}
-            onChange={this.handleChange('description')}
+            value={values.description}
+            onChange={handleChange('description')}
             className={classes.textField}
             margin="normal"
           /><br/> {
-            this.state.error && (<Typography component="p" color="error">
+            values.error && (<Typography component="p" color="error">
               <Icon color="error" className={classes.error}>error</Icon>
-              {this.state.error}</Typography>)
+              {values.error}</Typography>)
           }
         </CardContent>
         <CardActions>
-          <Button color="primary" variant="raised" onClick={this.clickSubmit} className={classes.submit}>Submit</Button>
-          <Link to='/seller/shops' className={classes.submit}><Button variant="raised">Cancel</Button></Link>
+          <Button color="primary" variant="contained" onClick={clickSubmit} className={classes.submit}>Submit</Button>
+          <Link to='/seller/shops' className={classes.submit}><Button variant="contained">Cancel</Button></Link>
         </CardActions>
       </Card>
     </div>)
-  }
 }
-
-NewShop.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(NewShop)
