@@ -1,30 +1,30 @@
-import React, {Component} from 'react'
-import Card, {CardContent} from 'material-ui/Card'
-import Typography from 'material-ui/Typography'
-import Avatar from 'material-ui/Avatar'
-import Grid from 'material-ui/Grid'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import React, {useState, useEffect} from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import Card from '@material-ui/core/Card'
+import CardContent from '@material-ui/core/CardContent'
+import Typography from '@material-ui/core/Typography'
+import Avatar from '@material-ui/core/Avatar'
+import Grid from '@material-ui/core/Grid'
 import {read} from './api-shop.js'
 import Products from './../product/Products'
 import {listByShop} from './../product/api-product.js'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     flexGrow: 1,
     margin: 30,
   },
   card: {
     textAlign: 'center',
-    paddingBottom: theme.spacing.unit * 2
+    paddingBottom: theme.spacing(2)
   },
   title: {
-    margin: theme.spacing.unit * 2,
+    margin: theme.spacing(2),
     color: theme.palette.protectedTitle,
     fontSize: '1.2em'
   },
   subheading: {
-    marginTop: theme.spacing.unit,
+    marginTop: theme.spacing(1),
     color: theme.palette.openTitle
   },
   bigAvatar: {
@@ -33,65 +33,82 @@ const styles = theme => ({
     margin: 'auto'
   },
   productTitle: {
-    padding:`${theme.spacing.unit * 3}px ${theme.spacing.unit * 2.5}px ${theme.spacing.unit}px ${theme.spacing.unit * 2}px`,
+    padding:`${theme.spacing(3)}px ${theme.spacing(2.5)}px ${theme.spacing(1)}px ${theme.spacing(2)}px`,
     color: theme.palette.openTitle,
     width: '100%',
     fontSize: '1.2em'
   }
-})
+}))
 
-class Shop extends Component {
-  constructor({match}) {
-    super()
-    this.state = {
-      shop: '',
-      products:[]
-    }
-    this.match = match
-  }
+export default function Shop({match}) {
+  const classes = useStyles();
+  const [shop, setShop] = useState('')
+  const [products, setProducts] = useState([])
+  const [error, setError] = useState('')
 
-  loadProducts = () => {
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
     listByShop({
-      shopId: this.match.params.shopId
-    }).then((data)=>{
+      shopId: match.params.shopId
+    }, signal).then((data)=>{
       if (data.error) {
-        this.setState({error: data.error})
+        setError(data.error)
       } else {
-        this.setState({products: data})
+        setProducts(data)
       }
     })
-  }
-
-  componentDidMount = () => {
-    this.loadProducts()
     read({
-      shopId: this.match.params.shopId
-    }).then((data) => {
+      shopId: match.params.shopId
+    }, signal).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setError(data.error)
       } else {
-        this.setState({shop: data})
+        setShop(data)
       }
     })
-  }
 
-  render() {
-    const logoUrl = this.state.shop._id
-          ? `/api/shops/logo/${this.state.shop._id}?${new Date().getTime()}`
+    return function cleanup(){
+      abortController.abort()
+    }
+
+  }, [match.params.shopId])
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
+    listByShop({
+      shopId: match.params.shopId
+    }, signal).then((data)=>{
+      if (data.error) {
+        setError(data.error)
+      } else {
+        setProducts(data)
+      }
+    })
+
+    return function cleanup(){
+      abortController.abort()
+    }
+
+  }, [match.params.shopId])
+
+    const logoUrl = shop._id
+          ? `/api/shops/logo/${shop._id}?${new Date().getTime()}`
           : '/api/shops/defaultphoto'
-    const {classes} = this.props
     return (<div className={classes.root}>
-      <Grid container spacing={24}>
+      <Grid container spacing={8}>
         <Grid item xs={4} sm={4}>
           <Card className={classes.card}>
             <CardContent>
               <Typography type="headline" component="h2" className={classes.title}>
-                {this.state.shop.name}
+                {shop.name}
               </Typography>
               <br/>
               <Avatar src={logoUrl} className={classes.bigAvatar}/><br/>
                 <Typography type="subheading" component="h2" className={classes.subheading}>
-                  {this.state.shop.description}
+                  {shop.description}
                 </Typography><br/>
             </CardContent>
           </Card>
@@ -99,16 +116,9 @@ class Shop extends Component {
         <Grid item xs={8} sm={8}>
           <Card>
             <Typography type="title" component="h2" className={classes.productTitle}>Products</Typography>
-            <Products products={this.state.products} searched={false}/>
+            <Products products={products} searched={false}/>
           </Card>
         </Grid>
       </Grid>
     </div>)
-  }
 }
-
-Shop.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(Shop)
