@@ -15,7 +15,7 @@ const auth = {
       sessionStorage.setItem('jwt', JSON.stringify(jwt))
     cb()
   },
-  signout(cb) {
+  clearJWT(cb) {
     if (typeof window !== "undefined")
       sessionStorage.removeItem('jwt')
     cb()
@@ -27,11 +27,11 @@ const auth = {
   updateUser(user, cb) {
     if(typeof window !== "undefined"){
       if(sessionStorage.getItem('jwt')){
-        let auth = JSON.parse(sessionStorage.getItem('jwt'))
-        auth.user = user
-        sessionStorage.setItem('jwt', JSON.stringify(auth))
-        cb()
-      }
+         let auth = JSON.parse(sessionStorage.getItem('jwt'))
+         auth.user = user
+         sessionStorage.setItem('jwt', JSON.stringify(auth))
+         cb()
+       }
     }
   }
 }
