@@ -1,108 +1,125 @@
-const create = (params, credentials, order, token) => {
-  return fetch('/api/orders/'+params.userId, {
-      method: 'POST',
+const create = async (params, credentials, order, token) => {
+  try {
+    let response = await fetch('/api/orders/'+params.userId, {
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json',
+          'Authorization': 'Bearer ' + credentials.t
+        },
+        body: JSON.stringify({order: order, token:token})
+      })
+      return response.json()
+    }catch(err) {
+      console.log(err)
+    }
+}
+
+const listByShop = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/orders/shop/'+params.shopId, {
+      method: 'GET',
+      signal: signal,
+      headers: {
+        'Accept': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
+    return response.json()
+  }catch(err){
+    console.log(err)
+  }
+}
+
+const update = async (params, credentials, product) => {
+  try {
+    let response = await fetch('/api/order/status/' + params.shopId, {
+      method: 'PUT',
       headers: {
         'Accept': 'application/json',
         'Content-Type': 'application/json',
         'Authorization': 'Bearer ' + credentials.t
       },
-      body: JSON.stringify({order: order, token:token})
+      body: JSON.stringify(product)
     })
-    .then((response) => {
-      return response.json()
-    }).catch((err) => console.log(err))
-}
-
-const listByShop = (params, credentials) => {
-  return fetch('/api/orders/shop/'+params.shopId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
     return response.json()
-  }).catch((err) => {
+  } catch(err){
     console.log(err)
-  })
+  }
 }
 
-const update = (params, credentials, product) => {
-  return fetch('/api/order/status/' + params.shopId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify(product)
-  }).then((response) => {
+const cancelProduct = async (params, credentials, product) => {
+  try {
+    let response = await fetch('/api/order/'+params.shopId+'/cancel/'+params.productId, {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: JSON.stringify(product)
+    })
     return response.json()
-  }).catch((err) => {
+  }catch(err){
     console.log(err)
-  })
+  }
 }
 
-const cancelProduct = (params, credentials, product) => {
-  return fetch('/api/order/'+params.shopId+'/cancel/'+params.productId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify(product)
-  }).then((response) => {
+const processCharge = async (params, credentials, product) => {
+  try {
+    let response = await fetch('/api/order/'+params.orderId+'/charge/'+params.userId+'/'+params.shopId, {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: JSON.stringify(product)
+    })
     return response.json()
-  }).catch((err) => {
+  } catch(err) {
     console.log(err)
-  })
+  }
 }
 
-const processCharge = (params, credentials, product) => {
-  return fetch('/api/order/'+params.orderId+'/charge/'+params.userId+'/'+params.shopId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify(product)
-  }).then((response) => {
+const getStatusValues = async (signal) => {
+  try {
+    let response = await fetch('/api/order/status_values', {
+      method: 'GET',
+      signal: signal
+    })
     return response.json()
-  }).catch((err) => {
+  }catch(err) { 
     console.log(err)
-  })
+  }
 }
 
-const getStatusValues = () => {
-  return fetch('/api/order/status_values', {
-    method: 'GET'
-  }).then((response) => {
+const listByUser = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/orders/user/'+params.userId, {
+      method: 'GET',
+      signal: signal,
+      headers: {
+        'Accept': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
     return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const listByUser = (params, credentials) => {
-  return fetch('/api/orders/user/'+params.userId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => {
+  }catch(err) {
     console.log(err)
-  })
+  }
 }
 
-const read = (params, credentials) => {
-  return fetch('/api/order/' + params.orderId, {
-    method: 'GET'
-  }).then((response) => {
+const read = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/order/' + params.orderId, {
+      method: 'GET',
+      signal: signal
+    })
     return response.json()
-  }).catch((err) => console.log(err))
+  } catch(err) {
+    console.log(err)
+  }
 }
 
 export {
