@@ -1,12 +1,14 @@
-import React, {Component} from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Typography from 'material-ui/Typography'
-import GridList, { GridListTile, GridListTileBar } from 'material-ui/GridList'
+import {makeStyles} from '@material-ui/core/styles'
+import Typography from '@material-ui/core/Typography'
+import GridList from '@material-ui/core/GridList'
+import GridListTile from '@material-ui/core/GridListTile'
+import GridListTileBar from '@material-ui/core/GridListTileBar'
 import {Link} from 'react-router-dom'
 import AddToCart from './../cart/AddToCart'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     display: 'flex',
     flexWrap: 'wrap',
@@ -26,7 +28,7 @@ const styles = theme => ({
     padding: '16px 0 10px'
   },
   title: {
-    padding:`${theme.spacing.unit * 3}px ${theme.spacing.unit * 2.5}px ${theme.spacing.unit * 2}px`,
+    padding:`${theme.spacing(3)}px ${theme.spacing(2.5)}px ${theme.spacing(2)}px`,
     color: theme.palette.openTitle,
     width: '100%'
   },
@@ -46,16 +48,16 @@ const styles = theme => ({
     color:'rgb(189, 222, 219)',
     display:'block'
   }
-})
-class Products extends Component {
-  render() {
-    const {classes} = this.props
+}))
+
+export default function Products(props){
+  const classes = useStyles()
     return (
       <div className={classes.root}>
-      {this.props.products.length > 0 ?
+      {props.products.length > 0 ?
         (<div className={classes.container}>
           <GridList cellHeight={200} className={classes.gridList} cols={3}>
-          {this.props.products.map((product, i) => (
+          {props.products.map((product, i) => (
             <GridListTile key={i} className={classes.tile}>
               <Link to={"/product/"+product._id}><img className={classes.image} src={'/api/product/image/'+product._id} alt={product.name} /></Link>
               <GridListTileBar className={classes.tileBar}
@@ -67,14 +69,10 @@ class Products extends Component {
               />
             </GridListTile>
           ))}
-        </GridList></div>) : this.props.searched && (<Typography type="subheading" component="h4" className={classes.title}>No products found! :(</Typography>)}
+        </GridList></div>) : props.searched && (<Typography variant="subheading" component="h4" className={classes.title}>No products found! :(</Typography>)}
       </div>)
-  }
 }
 Products.propTypes = {
-  classes: PropTypes.object.isRequired,
   products: PropTypes.array.isRequired,
   searched: PropTypes.bool.isRequired
 }
-
-export default withStyles(styles)(Products)
