@@ -1,5 +1,6 @@
-const create = (params, credentials, shop) => {
-  return fetch('/api/shops/by/'+ params.userId, {
+const create = async (params, credentials, shop) => {
+  try {
+    let response = await fetch('/api/shops/by/'+ params.userId, {
       method: 'POST',
       headers: {
         'Accept': 'application/json',
@@ -7,69 +8,82 @@ const create = (params, credentials, shop) => {
       },
       body: shop
     })
-    .then((response) => {
       return response.json()
-    }).catch((err) => console.log(err))
-}
-
-const list = () => {
-  return fetch('/api/shops', {
-    method: 'GET',
-  }).then(response => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const listByOwner = (params, credentials) => {
-  return fetch('/api/shops/by/'+params.userId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
+    } catch(err) { 
+      console.log(err)
     }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => {
-    console.log(err)
-  })
 }
 
-const read = (params, credentials) => {
-  return fetch('/api/shop/' + params.shopId, {
-    method: 'GET'
-  }).then((response) => {
+const list = async (signal) => {
+  try {
+    let response = await fetch('/api/shops', {
+      method: 'GET',
+      signal: signal
+    })
     return response.json()
-  }).catch((err) => console.log(err))
+  }catch(err) {
+    console.log(err)
+  }
 }
 
-const update = (params, credentials, shop) => {
-  return fetch('/api/shops/' + params.shopId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: shop
-  }).then((response) => {
+const listByOwner = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/shops/by/'+params.userId, {
+      method: 'GET',
+      signal: signal,
+      headers: {
+        'Accept': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
     return response.json()
-  }).catch((err) => {
+  }catch(err){
     console.log(err)
-  })
+  }
 }
 
-const remove = (params, credentials) => {
-  return fetch('/api/shops/' + params.shopId, {
-    method: 'DELETE',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
+const read = async (params, signal) => {
+  try {
+    let response = await fetch('/api/shop/' + params.shopId, {
+      method: 'GET',
+      signal: signal,
+    })
     return response.json()
-  }).catch((err) => {
+  }catch(err) {
     console.log(err)
-  })
+  }
+}
+
+const update = async (params, credentials, shop) => {
+  try {
+    let response = await fetch('/api/shops/' + params.shopId, {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: shop
+    })
+    return response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const remove = async (params, credentials) => {
+  try {
+    let response = await fetch('/api/shops/' + params.shopId, {
+      method: 'DELETE',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
+    return response.json()
+  } catch(err) {
+    console.log(err)
+  }
 }
 
 export {
