@@ -1,80 +1,98 @@
-import config from './../../config/config'
-const create = (user) => {
-  return fetch('/api/users/', {
-      method: 'POST',
+
+const create = async (user) => {
+  try {
+      let response = await fetch('/api/users/', {
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(user)
+      })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const list = async (signal) => {
+  try {
+    let response = await fetch('/api/users/', {
+      method: 'GET',
+      signal: signal,
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const read = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'GET',
+      signal: signal,
       headers: {
         'Accept': 'application/json',
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const update = async (params, credentials, user) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
       },
       body: JSON.stringify(user)
     })
-    .then((response) => {
-      return response.json()
-    }).catch((err) => console.log(err))
-}
-
-const list = () => {
-  return fetch('/api/users/', {
-    method: 'GET',
-  }).then(response => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const read = (params, credentials) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const update = (params, credentials, user) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify(user)
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const remove = (params, credentials) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'DELETE',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const stripeUpdate = (params, credentials, auth_code) => {
-  return fetch('/api/stripe_auth/'+params.userId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify({stripe: auth_code})
-  }).then((response)=> {
-    return response.json()
-  }).catch((err) => {
+    return await response.json()
+  } catch(err) {
     console.log(err)
-  })
+  }
+}
+
+const remove = async (params, credentials) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'DELETE',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const stripeUpdate = async (params, credentials, auth_code, signal) => {
+  try {
+    let response = await fetch ('/api/stripe_auth/'+params.userId, {
+      method: 'PUT',
+      signal: signal,
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: JSON.stringify({stripe: auth_code})
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
 }
 
 export {
