@@ -3,36 +3,40 @@ import jwt from 'jsonwebtoken'
 import expressJwt from 'express-jwt'
 import config from './../../config/config'
 
-const signin = (req, res) => {
-  User.findOne({
-    "email": req.body.email
-  }, (err, user) => {
-
-    if (err || !user)
-      return res.status('401').json({
-        error: "User not found"
+const signin = async (req, res) => {
+  try {
+    let user = await User.findOne({
+        "email": req.body.email
       })
 
-    if (!user.authenticate(req.body.password)) {
-      return res.status('401').send({
-        error: "Email and password don't match."
+      if (!user)
+        return res.status('401').json({
+          error: "User not found"
+        })
+
+      if (!user.authenticate(req.body.password)) {
+        return res.status('401').send({
+          error: "Email and password don't match."
+        })
+      }
+
+      const token = jwt.sign({
+        _id: user._id
+      }, config.jwtSecret)
+
+      res.cookie("t", token, {
+        expire: new Date() + 9999
       })
-    }
 
-    const token = jwt.sign({
-      _id: user._id
-    }, config.jwtSecret)
-
-    res.cookie("t", token, {
-      expire: new Date() + 9999
+      return res.json({
+        token,
+        user: {_id: user._id, name: user.name, email: user.email, seller: user.seller}
+      })
+  } catch (err) {
+    return res.status('401').json({
+      error: "Could not sign in"
     })
-
-    return res.json({
-      token,
-      user: {_id: user._id, name: user.name, email: user.email, seller: user.seller}
-    })
-
-  })
+  }
 }
 
 const signout = (req, res) => {
