@@ -1,14 +1,14 @@
 import Product from '../models/product.model'
-import _ from 'lodash'
+import extend from 'lodash/extend'
 import errorHandler from './../helpers/dbErrorHandler'
 import formidable from 'formidable'
 import fs from 'fs'
-import profileImage from './../../client/assets/images/profile-pic.png'
+import defaultImage from './../../client/assets/images/default.png'
 
 const create = (req, res, next) => {
   let form = new formidable.IncomingForm()
   form.keepExtensions = true
-  form.parse(req, (err, fields, files) => {
+  form.parse(req, async (err, fields, files) => {
     if (err) {
       return res.status(400).json({
         message: "Image could not be uploaded"
@@ -20,26 +20,31 @@ const create = (req, res, next) => {
       product.image.data = fs.readFileSync(files.image.path)
       product.image.contentType = files.image.type
     }
-    product.save((err, result) => {
-      if (err) {
-        return res.status(400).json({
-          error: errorHandler.getErrorMessage(err)
-        })
-      }
+    try {
+      let result = await product.save()
       res.json(result)
-    })
+    } catch (err){
+      return res.status(400).json({
+        error: errorHandler.getErrorMessage(err)
+      })
+    }
   })
 }
 
-const productByID = (req, res, next, id) => {
-  Product.findById(id).populate('shop', '_id name').exec((err, product) => {
-    if (err || !product)
+const productByID = async (req, res, next, id) => {
+  try {
+    let product = await Product.findById(id).populate('shop', '_id name').exec()
+    if (!product)
       return res.status('400').json({
         error: "Product not found"
       })
     req.product = product
     next()
-  })
+  } catch (err) {
+    return res.status('400').json({
+      error: "Could not retrieve product"
+    })
+  }
 }
 
 const photo = (req, res, next) => {
@@ -50,7 +55,7 @@ const photo = (req, res, next) => {
   next()
 }
 const defaultPhoto = (req, res) => {
-  return res.sendFile(process.cwd()+profileImage)
+  return res.sendFile(process.cwd()+defaultImage)
 }
 
 const read = (req, res) => {
@@ -58,106 +63,107 @@ const read = (req, res) => {
   return res.json(req.product)
 }
 
-const update = (req, res, next) => {
+const update = (req, res) => {
   let form = new formidable.IncomingForm()
   form.keepExtensions = true
-  form.parse(req, (err, fields, files) => {
+  form.parse(req, async (err, fields, files) => {
     if (err) {
       return res.status(400).json({
         message: "Photo could not be uploaded"
       })
     }
     let product = req.product
-    product = _.extend(product, fields)
+    product = extend(product, fields)
     product.updated = Date.now()
     if(files.image){
       product.image.data = fs.readFileSync(files.image.path)
       product.image.contentType = files.image.type
     }
-    product.save((err, result) => {
-      if (err) {
-        return res.status(400).send({
-          error: errorHandler.getErrorMessage(err)
-        })
-      }
+    try {
+      let result = await product.save()
       res.json(result)
-    })
-  })
-}
-
-const remove = (req, res, next) => {
-  let product = req.product
-  product.remove((err, deletedProduct) => {
-    if (err) {
+    }catch (err){
       return res.status(400).json({
         error: errorHandler.getErrorMessage(err)
       })
     }
+  })
+}
+
+const remove = async (req, res) => {
+  try{
+    let product = req.product
+    let deletedProduct = await product.remove()
     res.json(deletedProduct)
-  })
+  
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const listByShop = (req, res) => {
-  Product.find({shop: req.shop._id}, (err, products) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const listByShop = async (req, res) => {
+  try {
+    let products = await Product.find({shop: req.shop._id}).populate('shop', '_id name').select('-image')
     res.json(products)
-  }).populate('shop', '_id name').select('-image')
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const listLatest = (req, res) => {
-  Product.find({}).sort('-created').limit(5).populate('shop', '_id name').exec((err, products) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const listLatest = async (req, res) => {
+  try {
+    let products = await Product.find({}).sort('-created').limit(5).populate('shop', '_id name').exec()
     res.json(products)
-  })
+  } catch (err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const listRelated = (req, res) => {
-  Product.find({ "_id": { "$ne": req.product }, "category": req.product.category}).limit(5).populate('shop', '_id name').exec((err, products) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const listRelated = async (req, res) => {
+  try{
+    let products = await Product.find({ "_id": { "$ne": req.product }, "category": req.product.category}).limit(5).populate('shop', '_id name').exec()
     res.json(products)
-  })
+  } catch (err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const listCategories = (req, res) => {
-  Product.distinct('category',{},(err, products) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const listCategories = async (req, res) => {
+  try {
+    let products = await Product.distinct('category',{})
     res.json(products)
-  })
+  } catch (err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const list = (req, res) => {
+const list = async (req, res) => {
   const query = {}
   if(req.query.search)
     query.name = {'$regex': req.query.search, '$options': "i"}
   if(req.query.category && req.query.category != 'All')
     query.category =  req.query.category
-  Product.find(query, (err, products) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+  try {
+    let products = await Product.find(query).populate('shop', '_id name').select('-image').exec()
     res.json(products)
-  }).populate('shop', '_id name').select('-image')
+  } catch (err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const decreaseQuantity = (req, res, next) => {
+const decreaseQuantity = async (req, res, next) => {
   let bulkOps = req.body.order.products.map((item) => {
     return {
         "updateOne": {
@@ -166,26 +172,26 @@ const decreaseQuantity = (req, res, next) => {
         }
     }
    })
-   Product.bulkWrite(bulkOps, {}, (err, products) => {
-     if(err){
-       return res.status(400).json({
-         error: "Could not update product"
-       })
-     }
+   try {
+     await Product.bulkWrite(bulkOps, {})
      next()
-   })
+   } catch (err){
+      return res.status(400).json({
+        error: "Could not update product"
+      })
+   }
 }
 
-const increaseQuantity = (req, res, next) => {
-  Product.findByIdAndUpdate(req.product._id, {$inc: {"quantity": req.body.quantity}}, {new: true})
-    .exec((err, result) => {
-      if (err) {
-        return res.status(400).json({
-          error: errorHandler.getErrorMessage(err)
-        })
-      }
+const increaseQuantity = async (req, res, next) => {
+  try {
+    await Product.findByIdAndUpdate(req.product._id, {$inc: {"quantity": req.body.quantity}}, {new: true})
+    .exec()
       next()
+  } catch (err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
     })
+  }
 }
 
 export default {
