@@ -1,5 +1,5 @@
 import User from '../models/user.model'
-import _ from 'lodash'
+import extend from 'lodash/extend'
 import errorHandler from './../helpers/dbErrorHandler'
 import request from 'request'
 import config from './../../config/config'
@@ -7,32 +7,37 @@ import stripe from 'stripe'
 
 const myStripe = stripe(config.stripe_test_secret_key)
 
-const create = (req, res, next) => {
+const create = async (req, res) => {
   const user = new User(req.body)
-  user.save((err, result) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
-    res.status(200).json({
+  try {
+    await user.save()
+    return res.status(200).json({
       message: "Successfully signed up!"
     })
-  })
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
 /**
  * Load user and append to req.
  */
-const userByID = (req, res, next, id) => {
-  User.findById(id).exec((err, user) => {
-    if (err || !user)
+const userByID = async (req, res, next, id) => {
+  try {
+    let user = await User.findById(id)
+    if (!user)
       return res.status('400').json({
         error: "User not found"
       })
     req.profile = user
     next()
-  })
+  } catch (err) {
+    return res.status('400').json({
+      error: "Could not retrieve user"
+    })
+  }
 }
 
 const read = (req, res) => {
@@ -41,45 +46,45 @@ const read = (req, res) => {
   return res.json(req.profile)
 }
 
-const list = (req, res) => {
-  User.find((err, users) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const list = async (req, res) => {
+  try {
+    let users = await User.find().select('name email updated created')
     res.json(users)
-  }).select('name email updated created')
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const update = (req, res, next) => {
-  let user = req.profile
-  user = _.extend(user, req.body)
-  user.updated = Date.now()
-  user.save((err) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const update = async (req, res) => {
+  try {
+    let user = req.profile
+    user = extend(user, req.body)
+    user.updated = Date.now()
+    await user.save()
     user.hashed_password = undefined
     user.salt = undefined
     res.json(user)
-  })
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const remove = (req, res, next) => {
-  let user = req.profile
-  user.remove((err, deletedUser) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const remove = async (req, res) => {
+  try {
+    let user = req.profile
+    let deletedUser = await user.remove()
     deletedUser.hashed_password = undefined
     deletedUser.salt = undefined
     res.json(deletedUser)
-  })
+  } catch (err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
 const isSeller = (req, res, next) => {
@@ -153,14 +158,14 @@ const createCharge = (req, res, next) => {
   myStripe.tokens.create({
     customer: req.order.payment_id,
   }, {
-    stripe_account: req.profile.stripe_seller.stripe_user_id,
+    stripeAccount: req.profile.stripe_seller.stripe_user_id,
   }).then((token) => {
       myStripe.charges.create({
         amount: req.body.amount * 100, //amount in cents
         currency: "usd",
         source: token.id,
       }, {
-        stripe_account: req.profile.stripe_seller.stripe_user_id,
+        stripeAccount: req.profile.stripe_seller.stripe_user_id,
       }).then((charge) => {
         next()
       })
