@@ -16,12 +16,10 @@ import orderRoutes from './routes/order.routes'
 import React from 'react'
 import ReactDOMServer from 'react-dom/server'
 import MainRouter from './../client/MainRouter'
-import StaticRouter from 'react-router-dom/StaticRouter'
+import { StaticRouter } from 'react-router-dom'
 
-import { SheetsRegistry } from 'react-jss/lib/jss'
-import JssProvider from 'react-jss/lib/JssProvider'
-import { MuiThemeProvider, createMuiTheme, createGenerateClassName } from 'material-ui/styles'
-import { blueGrey, lightGreen } from 'material-ui/colors'
+import { ServerStyleSheets, ThemeProvider } from '@material-ui/styles'
+import theme from './../client/theme'
 //end
 
 //comment out before building for production
@@ -53,41 +51,21 @@ app.use('/', productRoutes)
 app.use('/', orderRoutes)
 
 app.get('*', (req, res) => {
-   const sheetsRegistry = new SheetsRegistry()
-   const theme = createMuiTheme({
-     palette: {
-       primary: {
-       light: '#8eacbb',
-       main: '#607d8b',
-       dark: '#34515e',
-       contrastText: '#fff',
-     },
-     secondary: {
-       light: '#e7ff8c',
-       main: '#b2ff59',
-       dark: '#7ecb20',
-       contrastText: '#000',
-     },
-       openTitle: blueGrey['400'],
-       protectedTitle: lightGreen['400'],
-       type: 'light'
-     }
-   })
-   const generateClassName = createGenerateClassName()
-   const context = {}
-   const markup = ReactDOMServer.renderToString(
+  const sheets = new ServerStyleSheets()
+  const context = {}
+  const markup = ReactDOMServer.renderToString(
+    sheets.collect(
       <StaticRouter location={req.url} context={context}>
-         <JssProvider registry={sheetsRegistry} generateClassName={generateClassName}>
-            <MuiThemeProvider theme={theme} sheetsManager={new Map()}>
-              <MainRouter/>
-            </MuiThemeProvider>
-         </JssProvider>
+          <ThemeProvider theme={theme}>
+            <MainRouter/>
+          </ThemeProvider>
       </StaticRouter>
      )
+  )
     if (context.url) {
       return res.redirect(303, context.url)
     }
-    const css = sheetsRegistry.toString()
+    const css = sheets.toString()
     res.status(200).send(Template({
       markup: markup,
       css: css
@@ -98,6 +76,9 @@ app.get('*', (req, res) => {
 app.use((err, req, res, next) => {
   if (err.name === 'UnauthorizedError') {
     res.status(401).json({"error" : err.name + ": " + err.message})
+  }else if (err) {
+    res.status(400).json({"error" : err.name + ": " + err.message})
+    console.log(err)
   }
 })
 
